Move tests in api java package straight to the test/java/javaclasses/exlibris... Move test util api to javaclasses.exlibris.testdata
