+++ api-java/src/test/java/javaclasses/exlibris/c/aggregate/BookCommandTest.java
@@ -18,40 +18,38 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package io.spine.javaclasses.exlibris.c.aggregate.definition;
+package javaclasses.exlibris.c.aggregate;
 
 import com.google.protobuf.Message;
 import io.spine.client.TestActorRequestFactory;
 import io.spine.core.CommandEnvelope;
 import io.spine.server.aggregate.AggregateCommandTest;
 import javaclasses.exlibris.BookId;
-import javaclasses.exlibris.InventoryId;
 import javaclasses.exlibris.Isbn62;
-import javaclasses.exlibris.c.aggregate.InventoryAggregate;
 
 /**
- * @author Alexander Karpets
+ * @author Paul Ageyev
  */
-public class InventoryCommandTest<C extends Message> extends AggregateCommandTest<C, InventoryAggregate> {
+abstract class BookCommandTest<C extends Message> extends AggregateCommandTest<C, BookAggregate> {
+
     private final TestActorRequestFactory requestFactory =
             TestActorRequestFactory.newInstance(getClass());
 
-    InventoryAggregate aggregate;
-    InventoryId inventoryId;
+    BookAggregate aggregate;
+    BookId bookId;
 
-    private static InventoryId createBookId() {
+    private static BookId createBookId() {
 
-        return InventoryId.newBuilder()
-                          .setBookId(BookId.newBuilder()
-                                           .setIsbn62(Isbn62.newBuilder()
-                                                            .setValue("123456789")))
-                          .build();
+        return BookId.newBuilder()
+                     .setIsbn62(Isbn62.newBuilder()
+                                      .setValue("1"))
+                     .build();
     }
 
     @Override
-    protected InventoryAggregate createAggregate() {
-        inventoryId = createBookId();
-        return new InventoryAggregate(inventoryId);
+    protected BookAggregate createAggregate() {
+        bookId = createBookId();
+        return new BookAggregate(bookId);
     }
 
     @Override
