+++ api-java/src/test/java/javaclasses/exlibris/c/aggregate/MarkLoanOverdueCommandTest.java
@@ -18,7 +18,7 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package io.spine.javaclasses.exlibris.c.aggregate.definition;
+package javaclasses.exlibris.c.aggregate;
 
 import com.google.protobuf.Message;
 import javaclasses.exlibris.Inventory;
@@ -31,11 +31,13 @@ import javaclasses.exlibris.c.MarkLoanOverdue;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+
 import java.util.List;
-import static io.spine.javaclasses.exlibris.testdata.InventoryCommandFactory.appendInventoryInstance;
-import static io.spine.javaclasses.exlibris.testdata.InventoryCommandFactory.borrowBookInstance;
-import static io.spine.javaclasses.exlibris.testdata.InventoryCommandFactory.markLoanOverdue;
+
 import static io.spine.server.aggregate.AggregateMessageDispatcher.dispatchCommand;
+import static javaclasses.exlibris.testdata.InventoryCommandFactory.appendInventoryInstance;
+import static javaclasses.exlibris.testdata.InventoryCommandFactory.borrowBookInstance;
+import static javaclasses.exlibris.testdata.InventoryCommandFactory.markLoanOverdue;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertTrue;
