+++ api-java/src/test/java/javaclasses/exlibris/c/aggregate/RemoveBookCommandTest.java
@@ -18,11 +18,10 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package io.spine.javaclasses.exlibris.c.aggregate.definition;
+package javaclasses.exlibris.c.aggregate;
 
 import com.google.common.base.Throwables;
 import com.google.protobuf.Message;
-import io.spine.javaclasses.exlibris.testdata.BookCommandFactory;
 import javaclasses.exlibris.Book;
 import javaclasses.exlibris.BookId;
 import javaclasses.exlibris.Isbn62;
@@ -30,16 +29,19 @@ import javaclasses.exlibris.c.AddBook;
 import javaclasses.exlibris.c.BookRemoved;
 import javaclasses.exlibris.c.RemoveBook;
 import javaclasses.exlibris.c.rejection.CannotRemoveMissingBook;
+import javaclasses.exlibris.testdata.BookCommandFactory;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+
 import java.util.List;
-import static io.spine.javaclasses.exlibris.testdata.BookCommandFactory.createBookInstance;
-import static io.spine.javaclasses.exlibris.testdata.BookCommandFactory.librarianId;
-import static io.spine.javaclasses.exlibris.testdata.BookCommandFactory.removalReason;
-import static io.spine.javaclasses.exlibris.testdata.BookCommandFactory.removeBookInstance;
-import static io.spine.javaclasses.exlibris.testdata.BookCommandFactory.userId;
+
 import static io.spine.server.aggregate.AggregateMessageDispatcher.dispatchCommand;
+import static javaclasses.exlibris.testdata.BookCommandFactory.createBookInstance;
+import static javaclasses.exlibris.testdata.BookCommandFactory.librarianId;
+import static javaclasses.exlibris.testdata.BookCommandFactory.removalReason;
+import static javaclasses.exlibris.testdata.BookCommandFactory.removeBookInstance;
+import static javaclasses.exlibris.testdata.BookCommandFactory.userId;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertThrows;
