+++ testutil-api/src/main/java/javaclasses/exlibris/testdata/InventoryCommandFactory.java
@@ -18,7 +18,7 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package io.spine.javaclasses.exlibris.testdata;
+package javaclasses.exlibris.testdata;
 
 import io.spine.net.EmailAddress;
 import javaclasses.exlibris.BookId;
